Squash warning about int64_t -> double implicit cast.
Make the cast explicit.

+++ pdf/pdf_font3.c
@@ -290,7 +290,7 @@ int pdfi_read_type3_font(pdf_context *ctx, pdf_dict *font_dict, pdf_dict *stream
         code = pdfi_dict_knownget(ctx, font->FontDescriptor, "MissingWidth", &obj);
         if (code > 0) {
             if (obj->type == PDF_INT) {
-                font->MissingWidth = ((pdf_num *) obj)->value.i;
+                font->MissingWidth = (double)((pdf_num *) obj)->value.i;
             }
             else if (obj->type == PDF_REAL) {
                 font->MissingWidth = ((pdf_num *) obj)->value.d;
